divide the main script into subfolders for clarity
